# ExtBlockMover: fix corrupt ee_start_hi on extent relocation (>>32 shift no-op)
PatchExtentNodeStream wrote ee_start_hi as (ushort)(newStart >> 32) with newStart a
32-bit uint. C# masks a 32-bit shift count to 31, so '>> 32' is a no-op and ee_start_hi
received the LOW 16 bits of the new physical block instead of zero — turning a relocate
to block 598 into a wild physical start 0x0255_00000256 that e2fsck rejects and the kernel
cannot read. Compute the physical block and new start as 64-bit so the shift extracts the
real high half (0 for sub-16-TiB images). This unblocks genuine ext block relocation for
defrag and (future) relocation-shrink, which had to be scoped out because of this defect.

+ ExtBlockMoverFsckTests (OsIntegration): mkfs.ext4 an extent-mapped file, relocate it
  through the public IFilesystemBlockMover surface, then assert e2fsck -fn is clean and
  debugfs reads the file back byte-identical at the new blocks. Reproduces red pre-fix
  (e2fsck exit 4) and passes post-fix.

+++ FileSystems/FileSystem.Ext/ExtBlockMover.cs
@@ -272,9 +272,13 @@ public sealed class ExtBlockMover : IFilesystemBlockMover {
         if (off + 12 > node.Length) break;
         var startHi = BinaryPrimitives.ReadUInt16LittleEndian(node.AsSpan(off + 6));
         var startLo = BinaryPrimitives.ReadUInt32LittleEndian(node.AsSpan(off + 8));
-        var startBlock = (uint)(((long)startHi << 32) | startLo);
-        if (startBlock >= oldFirst && startBlock < oldFirst + (uint)blockCount) {
-          var newStart = newFirst + (startBlock - oldFirst);
+        // 48-bit physical block: keep it 64-bit so the ee_start_hi write below
+        // shifts correctly. A 32-bit value here would make `newStart >> 32` a
+        // no-op (C# masks the shift to 31), writing the LOW 16 bits into
+        // ee_start_hi and producing a wild physical start that fsck rejects.
+        var startBlock = ((long)startHi << 32) | startLo;
+        if (startBlock >= oldFirst && startBlock < (long)oldFirst + blockCount) {
+          var newStart = (long)newFirst + (startBlock - oldFirst);
           BinaryPrimitives.WriteUInt16LittleEndian(node.AsSpan(off + 6), (ushort)(newStart >> 32));
           BinaryPrimitives.WriteUInt32LittleEndian(node.AsSpan(off + 8), (uint)(newStart & 0xFFFFFFFF));
           changed = true;
